// apps/api/src/generated/kysely/valuations/EventType.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Represents the enum valuations.EventType */
enum EventType {
  FOUNDER_EQUITY_SPLIT = 'FOUNDER_EQUITY_SPLIT',
  INVESTMENT_ROUND = 'INVESTMENT_ROUND',
  SHARE_SPLIT = 'SHARE_SPLIT',
  SHARE_REVERSE_SPLIT = 'SHARE_REVERSE_SPLIT',
  SECONDARY_SALE = 'SECONDARY_SALE',
  DISTRIBUTION = 'DISTRIBUTION',
  DIVIDEND = 'DIVIDEND',
  MARKDOWN = 'MARKDOWN',
  FUND_DISTRIBUTION = 'FUND_DISTRIBUTION',
  FUND_CLOSE = 'FUND_CLOSE',
  SHARE_PRICE = 'SHARE_PRICE',
  LIQUIDATION = 'LIQUIDATION',
};

export default EventType;